fix warnings

+++ tcob/src/gfx/ui/widgets/Displays.cpp
@@ -91,7 +91,7 @@ void dot_matrix_display::on_update(milliseconds /* deltaTime */)
     auto const& dots {Dots()};
     _sortedDots.clear();
     for (usize idx {0}; idx < dots.count(); ++idx) {
-        _sortedDots[dots[idx]].emplace_back(idx % width, idx / width);
+        _sortedDots[dots[idx]].emplace_back(static_cast<i32>(idx % width), static_cast<i32>(idx / width));
     }
 }
